Processing and generation of pipelines is now parallel

+++ pySPACEOptimizer/optimizer/base_optimizer.py
@@ -42,6 +42,7 @@ class PySPACEOptimizer(object):
         else:
             self._best_result_file = "%s_best.yaml" % task["data_set_path"]
         self._logger = logging.getLogger("%s.%s" % (self.__class__.__module__, self.__class__.__name__))
+        self.__pipelines = []
 
     def store_best_result(self, best_pipeline, best_parameters):
         """
@@ -55,18 +56,23 @@ class PySPACEOptimizer(object):
             best_result_file.write(operation_spec["base_file"])
 
     def _generate_pipelines(self):
-        for pipeline in PipelineGenerator(self._task):
-            self._logger.debug("Testing Pipeline: %s", pipeline)
-            pipeline = Pipeline(configuration=self._task,
-                                node_chain=[self._create_node(node_name) for node_name in pipeline])
-            if self._task.get("restart_evaluation", False) and os.path.isdir(pipeline.base_result_dir):
-                # Delete the old values and start over again
-                try:
-                    shutil.rmtree(pipeline.base_result_dir)
-                    os.makedirs(pipeline.base_result_dir)
-                except OSError as exc:
-                    pipeline.get_logger().warning("Error while trying to remove the old data: %s", exc)
-            yield pipeline
+        if not self.__pipelines:
+            for pipeline in PipelineGenerator(self._task):
+                self._logger.debug("Testing Pipeline: %s", pipeline)
+                pipeline = Pipeline(configuration=self._task,
+                                    node_chain=[self._create_node(node_name) for node_name in pipeline])
+                if self._task.get("restart_evaluation", False) and os.path.isdir(pipeline.base_result_dir):
+                    # Delete the old values and start over again
+                    try:
+                        shutil.rmtree(pipeline.base_result_dir)
+                        os.makedirs(pipeline.base_result_dir)
+                    except OSError as exc:
+                        pipeline.get_logger().warning("Error while trying to remove the old data: %s", exc)
+                self.__pipelines.append(pipeline)
+                yield pipeline
+        else:
+            for pipeline in self.__pipelines:
+                yield pipeline
 
     @abc.abstractmethod
     def _create_node(self, node_name):
@@ -81,13 +87,11 @@ class PySPACEOptimizer(object):
         raise NotImplementedError()
 
     @abc.abstractmethod
-    def _do_optimization(self, pipelines, evaluations, pass_):
+    def _do_optimization(self, evaluations, pass_):
         """
-        Do a single evaluation step on all given `pipelines` using `evaluations` evaluations.
+        Do a single evaluation step using `evaluations` evaluations.
         The `pass_` parameter is used to determinate the current pass this optimizer is in.
 
-        :param pipelines: The pipelines to optimize
-        :type pipelines: list[Pipeline]
         :param evaluations: The number of evaluations to do.
         :type evaluations: int
         :param pass_: The pass this optimization is done in.
@@ -112,12 +116,9 @@ class PySPACEOptimizer(object):
         evaluations = self._task["evaluations_per_pass"]
         passes = self._task["passes"]
 
-        pipelines = [pipeline for pipeline in self._generate_pipelines()]
-
         for pass_ in range(1, passes + 1):
             self._logger.info("-" * 10 + " Optimization pass: %d / %d " % (pass_, passes) + "-" * 10)
-            result = self._do_optimization(pipelines=pipelines,
-                                           evaluations=evaluations,
+            result = self._do_optimization(evaluations=evaluations,
                                            pass_=pass_)
             if result[0] < best[0]:
                 self._logger.info("Pipeline '%r' with parameters '%s' selected as best", result[1], result[2])
